chore: inline connector action declaration
